+++ lib/converters/safe_converters.dart
@@ -5,7 +5,7 @@ import 'package:json_annotation/json_annotation.dart';
 class SafeInt implements JsonConverter<int, Object?> {
   const SafeInt();
   @override
-  int fromJson(Object? json) => JSafe.int_(json);
+  int fromJson(Object? json) => JSafe.integer(json);
   @override
   Object? toJson(int object) => object;
 }
@@ -13,7 +13,7 @@ class SafeInt implements JsonConverter<int, Object?> {
 class SafeString implements JsonConverter<String, Object?> {
   const SafeString();
   @override
-  String fromJson(Object? json) => JSafe.str(json);
+  String fromJson(Object? json) => JSafe.string(json);
   @override
   Object? toJson(String object) => object;
 }
@@ -21,7 +21,7 @@ class SafeString implements JsonConverter<String, Object?> {
 class SafeBool implements JsonConverter<bool, Object?> {
   const SafeBool();
   @override
-  bool fromJson(Object? json) => JSafe.bool_(json);
+  bool fromJson(Object? json) => JSafe.boolean(json);
   @override
   Object? toJson(bool object) => object;
 }
@@ -29,7 +29,7 @@ class SafeBool implements JsonConverter<bool, Object?> {
 class SafeDouble implements JsonConverter<double, Object?> {
   const SafeDouble();
   @override
-  double fromJson(Object? json) => JSafe.dbl(json);
+  double fromJson(Object? json) => JSafe.double_(json);
   @override
   Object? toJson(double object) => object;
 }
@@ -37,7 +37,7 @@ class SafeDouble implements JsonConverter<double, Object?> {
 class SafeDateTime implements JsonConverter<DateTime, Object?> {
   const SafeDateTime();
   @override
-  DateTime fromJson(Object? json) => JSafe.dt(json);
+  DateTime fromJson(Object? json) => JSafe.dateTime(json);
   @override
   Object? toJson(DateTime object) => object.toIso8601String();
 }
@@ -46,7 +46,7 @@ class SafeDateTime implements JsonConverter<DateTime, Object?> {
 class SafeIntN implements JsonConverter<int?, Object?> {
   const SafeIntN();
   @override
-  int? fromJson(Object? json) => JSafe.intN(json);
+  int? fromJson(Object? json) => JSafe.integerN(json);
   @override
   Object? toJson(int? object) => object;
 }
@@ -54,7 +54,7 @@ class SafeIntN implements JsonConverter<int?, Object?> {
 class SafeStringN implements JsonConverter<String?, Object?> {
   const SafeStringN();
   @override
-  String? fromJson(Object? json) => JSafe.strN(json);
+  String? fromJson(Object? json) => JSafe.stringN(json);
   @override
   Object? toJson(String? object) => object;
 }
@@ -62,7 +62,7 @@ class SafeStringN implements JsonConverter<String?, Object?> {
 class SafeBoolN implements JsonConverter<bool?, Object?> {
   const SafeBoolN();
   @override
-  bool? fromJson(Object? json) => JSafe.boolN(json);
+  bool? fromJson(Object? json) => JSafe.booleanN(json);
   @override
   Object? toJson(bool? object) => object;
 }
@@ -70,7 +70,7 @@ class SafeBoolN implements JsonConverter<bool?, Object?> {
 class SafeDoubleN implements JsonConverter<double?, Object?> {
   const SafeDoubleN();
   @override
-  double? fromJson(Object? json) => JSafe.dblN(json);
+  double? fromJson(Object? json) => JSafe.doubleN(json);
   @override
   Object? toJson(double? object) => object;
 }
@@ -78,7 +78,7 @@ class SafeDoubleN implements JsonConverter<double?, Object?> {
 class SafeDateTimeN implements JsonConverter<DateTime?, Object?> {
   const SafeDateTimeN();
   @override
-  DateTime? fromJson(Object? json) => JSafe.dtN(json);
+  DateTime? fromJson(Object? json) => JSafe.dateTimeN(json);
   @override
   Object? toJson(DateTime? object) => object?.toIso8601String();
 }